offensive-recon-d/requirements.txt
python-whois
dnspython
requests
rich
python-nmap